+++ backend/src/main/java/kakaobootcamp/backend/domains/email/EmailService.java
@@ -1,7 +1,9 @@
 package kakaobootcamp.backend.domains.email;
 
 import static kakaobootcamp.backend.common.exception.ErrorCode.*;
+import static kakaobootcamp.backend.domains.email.dto.EmailDTO.*;
 
+import java.util.Objects;
 import java.util.Random;
 
 import org.springframework.mail.SimpleMailMessage;
@@ -13,8 +15,11 @@ import kakaobootcamp.backend.common.exception.ApiException;
 import kakaobootcamp.backend.common.exception.ErrorCode;
 import kakaobootcamp.backend.domains.email.domain.EmailCode;
 import kakaobootcamp.backend.domains.email.domain.VerifiedEmail;
+import kakaobootcamp.backend.domains.email.dto.EmailDTO;
 import kakaobootcamp.backend.domains.email.repository.EmailCodeRepository;
 import kakaobootcamp.backend.domains.email.repository.VerifiedEmailRepository;
+import kakaobootcamp.backend.domains.member.MemberService;
+import kakaobootcamp.backend.domains.member.dto.MemberDTO;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -25,8 +30,49 @@ public class EmailService {
 	private final JavaMailSender emailSender;
 	private final VerifiedEmailRepository verifiedEmailRepository;
 	private final EmailCodeRepository emailCodeRepository;
+	private final MemberService memberService;
 
-	public void sendEmail(
+	private static final String EMAIL_TITLE = "ANT HELPER 이메일 인증 코드";
+	private static final String EMAIL_TEXT = "인증 코드는 %d 입니다.";
+
+	// 이메일 인증번호 보내기
+	@Transactional(rollbackFor = ApiException.class)
+	public void validateEmailAndSendEmailVerification(SendVerificationCodeRequest request) {
+		String email = request.getEmail();
+
+		memberService.validateDuplicatedEmail(email);
+
+		// 본문
+		Integer verificationCode = createVerificationCode();
+		String text = String.format(EMAIL_TEXT, verificationCode);
+
+		// 이메일 보내기
+		sendEmail(email, EMAIL_TITLE, text);
+
+		// 이메일 코드 저장
+		checkEmailCodeDuplicationAndSaveEmailCode(email, verificationCode);
+	}
+
+	// 이메일 인증 확인
+	@Transactional(rollbackFor = ApiException.class)
+	public boolean verityEmailCode(VerifyEmailCodeRequest request) {
+		String email = request.getEmail();
+		Integer requestCode = request.getCode();
+
+		// 이메일로 인증 코드 검색
+		EmailCode emailCode = emailCodeRepository.findByEmail(email)
+			.orElseThrow(() -> ApiException.from(ErrorCode.UNAUTHENTICATED_EMAIL));
+
+		// 인증 코드 비교 후 인증된 이메일 저장
+		if (Objects.equals(emailCode.getCode(), requestCode)) {
+			saveVerifiedEmail(email);
+			return true;
+		}
+		return false;
+	}
+
+	// 이메일 보내기
+	private void sendEmail(
 		String toEmail,
 		String title,
 		String text) {
@@ -55,7 +101,7 @@ public class EmailService {
 
 	// 인증된 이메일 저장
 	@Transactional
-	public void saveVerifiedEmail(String email) {
+	private void saveVerifiedEmail(String email) {
 		VerifiedEmail verifiedEmail = new VerifiedEmail(email);
 		verifiedEmailRepository.save(verifiedEmail);
 	}
@@ -75,7 +121,7 @@ public class EmailService {
 	}
 
 	// 인증 코드 생성
-	public Integer createVerificationCode() {
+	private Integer createVerificationCode() {
 		Random random = new Random();
 
 		return random.nextInt(900000) + 100000; // 100000부터 999999까지의 숫자 생성
@@ -83,7 +129,7 @@ public class EmailService {
 
 	// 이메일 코드 중복 확인 및 저장
 	@Transactional
-	public void checkEmailCodeDuplicationAndSaveEmailCode(String email, Integer verificationCode) {
+	private void checkEmailCodeDuplicationAndSaveEmailCode(String email, Integer verificationCode) {
 		EmailCode emailCode = emailCodeRepository.findByEmail(email).orElse(null);
 
 		// 1. 이미 저장된 emailCode가 있으면 code만 변경
